python_application/requirements.txt
psycopg2-binary>=2.9.0
torch>=1.12.0
torchvision>=0.13.0
Pillow>=9.0.0
